+++ docs/sources/_conf.py
@@ -156,8 +156,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'Logtalk APIs', 'Logtalk APIs',
-     author, 'Logtalk APIs', 'Logtalk core, library, tools, and contributions APIs.',
+    (master_doc, 'LogtalkAPIs-' + version, 'Logtalk APIs',
+     u'Paulo Moura', 'Logtalk APIs', 'Logtalk core, library, tools, and contributions APIs.',
      'Miscellaneous'),
 ]